Fixes for  
CBL-598
CBL-599
CBL-615
CBL-616

+++ java/TravelSample/src/main/java/com/couchbase/travelsample/ui/controller/BookingsController.java
@@ -19,6 +19,7 @@ import java.util.List;
 import java.util.logging.Logger;
 
 import javax.annotation.Nonnull;
+import javax.annotation.Nullable;
 import javax.inject.Inject;
 import javax.inject.Singleton;
 import javax.swing.DefaultListModel;
@@ -43,6 +44,8 @@ public final class BookingsController extends PageController {
     @Nonnull
     private final FlightsDao flightsDao;
 
+    private String sessionId;
+
     @Inject
     public BookingsController(@Nonnull Nav nav, @Nonnull DbManager localStore, @Nonnull FlightsDao flightsDao) {
         super(BookingsView.PAGE_NAME, nav, localStore);
@@ -57,17 +60,29 @@ public final class BookingsController extends PageController {
     public void selectHotel() { toPage(HotelSearchView.PAGE_NAME); }
 
     @Override
-    protected void onClose() { }
+    protected void onClose() {
+        final String id = sessionId;
+        sessionId = null;
+        localStore.endSession(id);
+        flightsModel.clear();
+    }
 
-    public void fetchBookedFlights() { flightsDao.getBookedFlights(this::updateFlights); }
+    public void fetchBookedFlights() {
+        if (sessionId != null) { throw new IllegalStateException("Session already exists: " + sessionId); }
+        sessionId = localStore.startSession();
+        flightsDao.getBookedFlights(sessionId, this::updateFlights);
+    }
 
-    public void deleteBooking(Flight flight) {
+    public void deleteBooking(@Nullable Flight flight) {
+        if (flight == null) { return; }
         flightsDao.deleteBookedFlight(flight);
         flightsModel.removeElement(flight);
     }
 
-    void updateFlights(List<Flight> flights) {
+    private void updateFlights(@Nullable List<Flight> flights) {
+        if (sessionId == null) { return; }
         flightsModel.clear();
+        if (flights == null) { return; }
         for (Flight flight : flights) { flightsModel.addElement(flight); }
     }
 }
